Add a method to stop subscriptions.
Currently there isn't any way to stop a subscription without closing the link altogether, this is problematic if you have multiple subscriptions and you want to be able to start and stop them without opening multiple sockets.

This PR adds an API to do this by exposing a 'stop' function on WebSocketLink, which takes the requestId.
The requestId was also not publicly available, so I'm also adding the RequestId context entry to the Response, it is up to the consumer to get the id from the response and use it to stop a subscription.

I added a unit test for this and I also tested it in my app which is under development.